Fixes for return computations
Mainly worked to implement the return charts in a better way (i.e.,
fixed the stupid computations).
This entailed the requirement of a boost library (specifically boost
math to get newton-raphson). Added other minima/root computations from
Brent and Zhang for other possible aspect finding situations. For a
multi-planet aspect pattern, we'll likely want to find minimum span,
rather than a root.
More work needs to be done to bring even the UI of the return chart up
to a good state. Essentially, it's a derived chart, and we need to allow
the user to specify what they're varying. For a return chart, this will
be time and place (i.e., the return chart around which dates and
relocated for where). For a relocation chart, this would be a simple
respecification of the locality but keeping the time from actually
changing. For a return chart, we could compute multiple options (instead
of just the nearest one or the first one we find). The return chart
could also allow specification of the harmonic (e.g., H4 solar is a
harmonic that captures the quartisolars, demi solar, and the basic
return).

A derived chart should be storable with some reference back to the
original chart to make it easier to recompute if there is some sort of
rectification. Possibly they could be attached to the original chart
object and chart datafile. Then the astro-database can have a
tree-widget to reflect the file sub-entries (i.e., the derived charts).
Moreover, we should be able to create links to other directories that
can then be reflected in the tree as well.

Also fixed the "default speed" of many bodies in the planets CSV table.
This was needed to get a decent step for the initial bracketing when
looking for return aspects. What we also need is a sidereal period
value, so that if we're searching for a synodic aspect between two
moving (or more) bodies, we can get a better idea of the range; again,
this is used to help bracket the algorithm-based searches like
newton-raphson and brent-zhang (and just the brent-minimum).