reimplement events lookup in main editor
Instead of having a special editor, make an 'events' type and expose event-lookup controls when that type is selected. Further, this allows finding
transit to natal events.

Fixed issue where closing earlier tab would cause crash.

Translate special file-system characters in chart names to %xx and back so that they can be in the file system. Especially important for charts of midpoint transits.