Updated SWEph; misc
Updated the Swiss ephemeris engine code to 2.09.
Use new API to get speed of Asc/MC to allow transit-to-transiting angle. Not sure it actually works that well, and it might be better to use the other API.
Fixed regression dateTimeFromJulian() which accidentally got checked in with wrong API.
Playing more with minimum spread finding: a crude bracketing mechanism to allow multiple minima.
Bracketing for newton-raphson lookup is still not great: easy to miss transits because of sensitivity to the time range.
Fixed (?) dage range saving/loading.
Enable hierarchical chart directories.
Disable qDebug() in release build, enable in debug build.