threading infra start + other fixes
fix transit of Asc/MC in tropical ayanamsha;
for now, address issue with spread calc to use  m=100 instead of m=1, the latter which was too small for catching some minima.
move transit finding functions to runnable classes. Tried threading the derived aspect finding, but that's going to require further refactor so that these are not dependent on data in the primary finder, which may change that data or go away before the sub-subthread is complete.
Derived aspects are now stored as a child list ("coincidences") of the original dated entry, so the event-to-model code changed accordingly.

tweaked the swiss ephemeris code to enable multi-threading on windows.