some midpoint transit robustness + aspect pattern finding
midpoint transits were not working great; possibly because the rate is
4 days per test. Some H8 searches without midpoints are also failing.
Perhaps the _rate can be adjusted depending on what is in the transit
list, e.g., if the moon or a moon+x midpoint are in the transit list, reduce the date accordingly. the highest harmonic itself can be a factor. Possibly also
the rate can be different for different harmonics, though I'm not quite
sure how that would work.

Also added a simplified clustering function. This is somewhat simpler than the harmonics dock widget code, and hopefully faster. It should be usable when the pattern-finding code is implemented. For now, it is able to find aspect patterns for stations and for returns. The spread-orb is currently hard-coded at 8 -- should it be less? The simplicity of the algorithm would be lost if different cluster-populations had different spread minima.

Also fixed problem with midpoint computation. Previously this was wrong, and caused the midpoint transits only to appear in H2 for certain planetary combinations.