Adaptive rate + click on transit harmonic to set harmonic
Playing with adaptive sample rate changes. Sort of works, but it might be better to try to converge with a smaller sample rate. Idea: generate statistics for successes vs. failures so that the typical convergence rate for the various algorithms can be profiles.

Some const correctness fixes for harmonics table generator.

Allow user to automatically set the harmonic display when clicking on the harmonics column for a particular row.