Various
focal aspect patterns for non-natal aspects;
fixed issue with detection of slow-moving planets (which is used to choose
the algorithm for convergence), which makes the newton-raphson
option get selected more readily.
fixed issue with date conversion so that fractional seconds can be used
for transit / ingress charts (charts would indicate ingressing planet in
last second last minute last degree of sign because the chart was off
by a fraction of a second).