Fixing problems with dubious result
Added check to see if cluster was truly
out of orb--discovered that a cluster
may not show up if another planet joins
the pattern within the range, which was effectively bumping the cluster
out of the set. This also led to the
repeated duplications.

Now the problem will be to gracefully determine when a cluster actually starts being in orb, and possibly also when it is going out of orb. This will be moderately cheap -- not all planets need to be calculated, however, it may be time to start exposing start and end
times so that it's clearer that a pattern
may have activity over a longer period
of time, rather than simply its local minimum.