Paired aspect timespan
and other miscellaneous tweaks to get correct behavior scrolling during.
Timespan stuff will be used to make some things more efficient but is not yet complete: for example, if we know that the minimum spread sampled across the timespan is outside of the range of search, we might choose not to find the precise minimum. Should save this information so it can be used in transits.cpp to annotate secondary aspect list (still work in progress).

Note still occasionally seeing disrupted/corrupted aspect list during scrolling. Need to track this down...